decimal

+++ public/js/situacion.js
@@ -17,7 +17,7 @@
 /*+++++++++++++++++++++++script sumatoria+++++++++++++++++++++++++++++++++++++++++++*/
     function sumar (valor) {
     var total = 0;
-    valor = parseInt(valor); // Convertir el valor a un entero (número).
+    valor = parseFloat(valor); // Convertir el valor a un entero (número).
 
     total = document.getElementById('spTotal').innerHTML;
 
@@ -25,7 +25,7 @@
     total = (total == null || total == undefined || total == "") ? 0 : total;
 
     /* Esta es la suma. */
-    total = (parseInt(total) + parseInt(valor));
+    total = (parseFloat(total) + parseFloat(valor));
 
     // Colocar el resultado de la suma en el control "span".
     document.getElementById('spTotal').innerHTML = total;
